Cập nhật
thêm các method mới, sửa lại refinegan, thêm các method onnx, đổi torch.float64 sang float32,....

+++ main/library/predictors/WORLD.py
@@ -1,12 +1,11 @@
 import os
-import shutil
 import torch
+import shutil
 import ctypes
 import platform
+
 import numpy as np
 import tempfile as tf
-
-
 
 model = torch.load(os.path.join("assets", "models", "predictors", "world.pth"), map_location="cpu")
 
@@ -21,6 +20,7 @@ with tf.NamedTemporaryFile(delete=False, suffix=suffix, dir=temp_folder) as temp
     temp_path = temp_file.name
 
 world_dll = ctypes.CDLL(temp_path)
+
 
 class DioOption(ctypes.Structure):
     _fields_ = [("F0Floor", ctypes.c_double), ("F0Ceil", ctypes.c_double), ("ChannelsInOctave", ctypes.c_double), ("FramePeriod", ctypes.c_double), ("Speed", ctypes.c_int), ("AllowedRange", ctypes.c_double)]
@@ -50,7 +50,7 @@ def harvest(x, fs, f0_floor=50, f0_ceil=1100, frame_period=10):
     tpos = (ctypes.c_double * f0_length)()
 
     world_dll.Harvest((ctypes.c_double * len(x))(*x), len(x), fs, ctypes.byref(option), tpos, f0)
-    return np.array(f0, dtype=np.float64), np.array(tpos, dtype=np.float64)
+    return np.array(f0, dtype=np.float32), np.array(tpos, dtype=np.float32)
 
 def dio(x, fs, f0_floor=50, f0_ceil=1100, channels_in_octave=2, frame_period=10, speed=1, allowed_range=0.1):
     world_dll.Dio.argtypes = [ctypes.POINTER(ctypes.c_double), ctypes.c_int, ctypes.c_int, ctypes.POINTER(DioOption), ctypes.POINTER(ctypes.c_double), ctypes.POINTER(ctypes.c_double)]
@@ -77,7 +77,7 @@ def dio(x, fs, f0_floor=50, f0_ceil=1100, channels_in_octave=2, frame_period=10,
     tpos = (ctypes.c_double * f0_length)()
 
     world_dll.Dio((ctypes.c_double * len(x))(*x), len(x), fs, ctypes.byref(option), tpos, f0)
-    return np.array(f0, dtype=np.float64), np.array(tpos, dtype=np.float64)
+    return np.array(f0, dtype=np.float32), np.array(tpos, dtype=np.float32)
 
 def stonemask(x, fs, tpos, f0):
     world_dll.StoneMask.argtypes = [ctypes.POINTER(ctypes.c_double), ctypes.c_int, ctypes.c_int, ctypes.POINTER(ctypes.c_double), ctypes.POINTER(ctypes.c_double), ctypes.c_int, ctypes.POINTER(ctypes.c_double)]
@@ -87,4 +87,4 @@ def stonemask(x, fs, tpos, f0):
     world_dll.StoneMask((ctypes.c_double * len(x))(*x), len(x), fs, (ctypes.c_double * len(tpos))(*tpos), (ctypes.c_double * len(f0))(*f0), len(f0), out_f0)
 
     if os.path.exists(temp_folder): shutil.rmtree(temp_folder, ignore_errors=True)
-    return np.array(out_f0, dtype=np.float64)
+    return np.array(out_f0, dtype=np.float32)
